Make sparsearrays an ext
Fixes https://github.com/JuliaArrays/ArrayInterface.jl/issues/447
