+++ Project.toml
@@ -5,8 +5,6 @@ version = "7.12.0"
 [deps]
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
-SuiteSparse = "4607b0f0-06f3-5cda-b6b1-a6196a1729e9"
 
 [weakdeps]
 BandedMatrices = "aae01518-5342-5314-be14-df237901396f"
@@ -16,6 +14,7 @@ CUDSS = "45b445bb-4962-46a0-9369-b4df9d0f772e"
 ChainRules = "082447d4-558c-5d27-93f4-14fc19e9eca2"
 GPUArraysCore = "46192b85-c4d5-4398-a991-12ede77f4527"
 ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
+SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StaticArraysCore = "1e83bf80-4336-4d27-bf5d-d5a4f845583c"
 Tracker = "9f7883ad-71c0-57eb-9f7f-b5c9e6d3789c"
 
@@ -27,6 +26,7 @@ ArrayInterfaceCUDSSExt = "CUDSS"
 ArrayInterfaceChainRulesExt = "ChainRules"
 ArrayInterfaceGPUArraysCoreExt = "GPUArraysCore"
 ArrayInterfaceReverseDiffExt = "ReverseDiff"
+ArrayInterfaceSparseArraysExt = "SparseArrays"
 ArrayInterfaceStaticArraysCoreExt = "StaticArraysCore"
 ArrayInterfaceTrackerExt = "Tracker"
 
@@ -42,7 +42,6 @@ LinearAlgebra = "1.10"
 ReverseDiff = "1"
 SparseArrays = "1.10"
 StaticArraysCore = "1"
-SuiteSparse = "1.10"
 Tracker = "0.2"
 julia = "1.10"
